# swagger.yaml
openapi: 3.0.0
info:
  title: Dynamic API
  version: 1.0.0
...
paths:
...
  /blog:
    post:
      tags:
        - blog
      summary: Insert new blog
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/blog'
      responses:
        '200':
          description: blog inserted successfully
    get:
      tags:
        - blog
      summary: Get all blogs (filtered)
      requestBody:
        content:
          application/json:
            schema:
              type: object
              description: Filter query
      responses:
        '200':
          description: List of blogs
    put:
      tags:
        - blog
      summary: Update existing blog
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/blog'
      responses:
        '200':
          description: blog updated successfully
    delete:
      tags:
        - blog
      summary: Delete blog by filter
      requestBody:
        content:
          application/json:
            schema:
              type: object
              description: Filter criteria for delete
      responses:
        '200':
          description: blog deleted successfully
  /blog/{id}:
    get:
      tags:
        - blog
      summary: Get blog by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: blog found
        '404':
          description: blog not found
...
  /faq:
    post:
      tags:
        - faq
      summary: Insert new faq
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/faq'
      responses:
        '200':
          description: faq inserted successfully
    get:
      tags:
        - faq
      summary: Get all faqs (filtered)
      requestBody:
        content:
          application/json:
            schema:
              type: object
              description: Filter query
      responses:
        '200':
          description: List of faqs
    put:
      tags:
        - faq
      summary: Update existing faq
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/faq'
      responses:
        '200':
          description: faq updated successfully
    delete:
      tags:
        - faq
      summary: Delete faq by filter
      requestBody:
        content:
          application/json:
            schema:
              type: object
              description: Filter criteria for delete
      responses:
        '200':
          description: faq deleted successfully
...
components:
  schemas:
...
    blog:
      type: object
      properties:
        blogId:
          type: string
        title:
          type: string
        subtitle:
          type: string
        slug:
          type: string
        tags:
          type: array
          items:
            type: string
        metaTitle:
          type: string
        metaDescription:
          type: string
        bannerdescription:
          type: string
        category:
          type: string
        author:
          type: string
        bannerImage:
          type: string
        thumbnailImage:
          type: string
        description:
          type: string
        createdOn:
          type: string
          format: date-time
        isActive:
          type: boolean
...
    faq:
      type: object
      properties:
        faqId:
          type: string
        title:
          type: string
        qa:
          type: array
          items:
            type: string
        isActive:
          type: boolean
...